Move old parsing code into `legacy/` dir
