+++ packages/sync-rules/src/index.ts
@@ -14,18 +14,11 @@ export * from './json_schema.js';
 export * from './request_functions.js';
 export * from './schema-generators/schema-generators.js';
 export * from './SourceTableRef.js';
-export * from './sql_filters.js';
 export * from './sql_functions.js';
-export * from './sql_support.js';
-export { SqlBucketDescriptor } from './SqlBucketDescriptor.js';
-export * from './SqlDataQuery.js';
-export * from './SqlParameterQuery.js';
+export * from './sqliteBool.js';
 export * from './SqlSyncRules.js';
 export * from './StaticSchema.js';
 export * from './StorageVersion.js';
-export { syncStreamFromSql } from './streams/from_sql.js';
-export { STREAM_FUNCTIONS } from './streams/functions.js';
-export { SyncStream } from './streams/stream.js';
 export { SyncConfig, SyncConfigWithErrors } from './SyncConfig.js';
 export * from './TablePattern.js';
 export * from './types.js';
