+++ packages/sync-rules/src/legacy/StaticSqlParameterQuery.ts
@@ -1,16 +1,17 @@
 import { SelectFromStatement } from 'pgsql-ast-parser';
-import { BucketDescription, BucketPriority, DEFAULT_BUCKET_PRIORITY } from './BucketDescription.js';
-import { BucketParameterQuerier, PendingQueriers } from './BucketParameterQuerier.js';
-import { CreateSourceParams } from './BucketSource.js';
-import { SqlRuleError } from './errors.js';
-import { BucketDataScope } from './HydrationState.js';
-import { BucketDataSource, BucketParameterQuerierSource, GetQuerierOptions, resolvedBucket } from './index.js';
-import { SourceTableRef } from './SourceTableRef.js';
+import { BucketDescription, BucketPriority, DEFAULT_BUCKET_PRIORITY } from '../BucketDescription.js';
+import { BucketParameterQuerier, PendingQueriers } from '../BucketParameterQuerier.js';
+import { CreateSourceParams } from '../BucketSource.js';
+import { SqlRuleError } from '../errors.js';
+import { BucketDataScope } from '../HydrationState.js';
+import { BucketDataSource, BucketParameterQuerierSource, GetQuerierOptions, resolvedBucket } from '../index.js';
+import { SourceTableRef } from '../SourceTableRef.js';
+import { sqliteBool } from '../sqliteBool.js';
+import { TablePattern } from '../TablePattern.js';
+import { ParameterValueClause, QueryParseOptions, RequestParameters, SqliteJsonValue } from '../types.js';
+import { bucketDescription, isJsonValue, serializeBucketParameters } from '../utils.js';
 import { AvailableTable, SqlTools } from './sql_filters.js';
-import { checkUnsupportedFeatures, isClauseError, sqliteBool } from './sql_support.js';
-import { TablePattern } from './TablePattern.js';
-import { ParameterValueClause, QueryParseOptions, RequestParameters, SqliteJsonValue } from './types.js';
-import { bucketDescription, isJsonValue, serializeBucketParameters } from './utils.js';
+import { checkUnsupportedFeatures, isClauseError } from './sql_support.js';
 import { DetectRequestParameters } from './validators.js';
 
 export interface StaticSqlParameterQueryOptions {
